Move the code to a dedicated package to be able to use it as a library
