+++ iothub/iothubHttp_test.go
@@ -1,4 +1,4 @@
-package main
+package iothub
 
 import (
 	"fmt"
